requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.6.5
appdirs==1.4.4
appnope==0.1.2
apturl==0.5.2
asgiref==3.4.1
astroid==2.4.1
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.2.0
beautifulsoup4==4.8.2
beepy==1.0.7
bitarray==1.2.2
black==21.11b1
blinker==1.4
bresenham==0.2.1
Brlapi==0.7.0
bs4==0.0.1
cached-property==1.5.2
cachelib==0.1.1
certifi==2022.6.15
cffi==1.14.2
chardet==3.0.4
charset-normalizer==2.0.9
chromedriver-autoinstaller==0.4.0
click==8.1.3
colorama==0.4.3
command-not-found==0.3
comtypes==1.1.14
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
cytoolz==0.11.2
dataclassy==0.11.1
dateparser==1.1.0
DateTime==4.3
dbus-python==1.2.16
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
demjson==2.2.4
discord==1.7.3
discord.py==1.7.3
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.7
django-crispy-forms==1.13.0
dnspython==2.0.0
duplicity==0.8.12.0
eip712==0.1.0
email-validator==1.1.1
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.2
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
fake-useragent==0.1.11
Faker==13.3.4
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Session==0.3.2
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-Table==0.5.0
flask-util-js==0.2.25
Flask-WTF==0.14.3
frozenlist==1.2.0
future==0.18.2
geographiclib==1.50
geopy==2.0.0
gevent==21.1.1
googlemaps==4.4.2
greenlet==1.0.0
gunicorn==20.1.0
h11==0.13.0
hexbytes==0.2.2
html5lib==1.0.1
httplib2==0.14.0
hyperlink==21.0.0
hypothesis==6.27.3
idna==2.10
importlib-metadata==4.12.0
incremental==21.3.0
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipython==7.30.1
isort==4.3.21
itsdangerous==2.0.0
jedi==0.18.1
Jinja2==3.0.0
joblib==1.1.0
jsonify==0.5
jsonschema==3.2.0
jtutils==0.0.8
kazam==1.5.3
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
leven==1.0.4
lockfile==0.12.2
louis==3.12.0
lru-dict==1.1.7
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
MechanicalSoup==1.1.0
monotonic==1.5
MouseInfo==0.1.3
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
netifaces==0.10.4
nltk==3.6.5
nose==1.3.7
numpy==1.19.0
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
opencv-python==4.3.0.36
outcome==1.2.0
packaging==21.3
panda==0.3.1
pandas==1.3.3
paramiko==2.6.0
parse==1.19.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pause==0.3
pbr==5.9.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.1.2
platformdirs==2.3.0
playsound==1.2.2
pluggy==1.0.0
prompt-toolkit==3.0.23
protobuf==3.6.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
PyAutoGUI==0.9.50
pycairo==1.16.2
pycparser==2.20
pycryptodome==3.12.0
pycups==1.9.73
pyee==8.2.2
pygame==2.0.1
PyGetWindow==0.0.8
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyGObject==3.36.0
PyJWT==2.4.0
pylint==2.5.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyMsgBox==1.0.8
PyNaCl==1.3.0
pyparsing==2.4.7
pyperclip==1.8.0
pyppeteer==1.0.2
pyquery==1.4.3
PyRect==0.1.4
pyRFC3339==1.1
pyrsistent==0.18.0
PyScreeze==0.1.26
PySocks==1.7.1
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-apt==2.0.0+ubuntu0.20.4.5
python-binance==1.0.15
python-csv==0.0.13
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.16.0
python-editor==1.0.4
python-engineio==3.13.2
python-math==0.0.1
python-socketio==4.6.0
python-vlc==3.0.11115
python-xlib==0.23
python3-xlib==0.15
pythx==1.6.1
PyTweening==1.0.3
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
random-user-agent==1.0.1
regex==2021.8.28
reportlab==3.5.34
requests==2.24.0
requests-html==0.10.0
requests-unixsocket==0.2.0
rlp==2.0.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.4.3
selenium-stealth==1.0.6
semantic-version==2.8.5
simpleaudio==1.0.4
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.0
SQLAlchemy==1.3.19
sqlparse==0.4.2
stripe==2.56.0
systemd-python==234
testresources==2.0.1
tk==0.1.0
toml==0.10.1
tomli==1.2.2
toolz==0.11.2
tqdm==4.62.3
traitlets==5.1.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==21.7.0
twitter-scraper==0.4.4
typed-ast==1.5.1
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
uiautomation==2.0.17
ujson==4.3.0
unattended-upgrades==0.1
undetected-chromedriver==3.1.5.post4
urllib3==1.26.12
usb-creator==0.3.7
varint==1.0.2
vboxapi==1.0
virtualenv==20.4.0
vvm==0.1.0
vyper==0.3.1
w3lib==1.22.0
wadllib==1.3.3
wcwidth==0.2.5
web3==5.25.0
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.57.0
websockets==9.1
Werkzeug==2.0.0
wrapt==1.12.1
wsproto==1.1.0
WTForms==2.3.3
xkit==0.0.0
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.7.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.1.0
